requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.1
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers==3.9.0

PyJWT==2.1.0
asgiref==3.4.1
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0